split parse func and change operator=None to dummy_operator
