+++ generate_dataset/modeling/__init__.py
@@ -1,4 +1,4 @@
 from .base_classes import BaseIndividual, BaseOperator, Term, Assertion, Formula, FACT_TYPE, FACT_T
-from .co_namespace import Declared_Operators
 from .topv2 import *
 from .conic10k import *
+from .co_namespace import Declared_Operators  # Declared_Operator必须得先被导入某个建模后，才可以使用
